Microprofile Metrics 2.0.0 in support of Microprofile 3.0 specification
I have added support for the base metrics within Jemo and written test
cases to cover 100% of the source code added. By definition of the
specification this now makes Jemo (basically compatible) with the
specification. That being said before moving forward with this branch I
want to add full support for application level metrics and the
OpenMetrics format.